# forms-flow-bpm/api-spec/forms-flow-bpm-api-spec.yaml
openapi: 3.0.0
info:
  title: forms-flow-bpm API
  description: The API for forms-flow-bpm.
  version: 1.0.0
servers:
  - url: 'http://{{baseurl}}'
  - url: 'http://{{baseurl}}/engine-rest'
  - url: 'http://localhost:8000'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    TenantAuthorization:
      title: TenantAuthorization
      type: object
      properties:
        tenantKey:
          type: string
        adminRoles:
          type: array
          items:
            type: string
        designerRoles:
          type: array
          items:
            type: string
        clientRoles:
          type: array
          items:
            type: string
        reviewerRoles:
          type: array
          items:
            type: string
      description: Tenant authorization model
security:
  - bearerAuth: []
...
paths:
...
  /engine-rest-ext/tenant/authorization:
    get:
      summary: Your GET endpoint
      tags: []
      responses: {}
      operationId: get-engine-rest-ext-tenant-authorization
    post:
      summary: ''
      operationId: post-engine-rest-ext-tenant-authorization
      responses:
        '200':
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantAuthorization'
            examples:
              example-1:
                value:
                  tenantKey: string
                  adminRoles:
                    - camunda-admin
                  designerRoles:
                    - formsflow-designer
                  clientRoles:
                    - formsflow-client
                  reviewerRoles:
                    - formsflow-reviewer